Simple entryscript to support 3 use-cases
This entrypoint should support 3 use-cases:

1. Untouched docker image for apache (CMD is apache2-xxx)
2. Untouched docker image for php-fpm (CMD is php-fpm)
3. Extended version with custom CMD

The 2 first use-cases are handled by looking a the parameter (the CMD command).

The last one is defined by the user when extending the image. He must
know whether he extends the apache or the php-fpm image and define the
NEXTCLOUD_DOCKER_MODE value with either "apache" or "php-fpm"

This PR also removes the NEXTCLOUD_UPDATE=1 that was used to force the
script to run when the image had been extended.